fix: extract email sign in util functions into files
