+++ client/src/app/authentication/utils/emailVerifCreateSchema.ts
@@ -1,0 +1,11 @@
+import * as yup from 'yup'
+
+export type EmailVerifCreateInput = {
+    email: string
+}
+
+const emailVerifCreateSchema = yup.object().shape({
+    email: yup.string().email().required()
+})
+
+export default emailVerifCreateSchema
